Use fixed-width int64_t + PRId64 for the touched log lines
Following review feedback, switch these spots to standard fixed-length
types (int64_t) printed via the PRId64 macro from <inttypes.h>, matching
the existing convention (config.c) and the goal of #4447, rather than
long long + %lld. Also drops the now-redundant (long long) casts.

+++ src/lvl_script_commands.c
@@ -12,6 +12,7 @@
  */
 /******************************************************************************/
 #include "pre_inc.h"
+#include <inttypes.h>
 #include <math.h>
 #include <string.h>
 #include "bflib_sound.h"
@@ -4891,7 +4892,7 @@ static void set_power_configuration_check(const struct ScriptLine *scline)
         DEALLOCATE_SCRIPT_VALUE
         return;
     }
-    long long number_value = 0;
+    int64_t number_value = 0;
     long k;
     switch (powervar)
     {
@@ -5127,15 +5128,15 @@ static void set_power_configuration_check(const struct ScriptLine *scline)
     {
         if ( (powervar == 5) && (value->chars[3] != -1) )
         {
-            SCRIPTDBG(7, "Toggling %s castability flag: %lld", powername, (long long)number_value);
+            SCRIPTDBG(7, "Toggling %s castability flag: %" PRId64, powername, number_value);
         }
         else if ( (powervar == 14) && (value->chars[3] != -1) )
         {
-            SCRIPTDBG(7, "Toggling %s property flag: %lld", powername, (long long)number_value);
+            SCRIPTDBG(7, "Toggling %s property flag: %" PRId64, powername, number_value);
         }
         else
         {
-            SCRIPTDBG(7, "Setting power %s property %s to %lld", powername, property, (long long)number_value);
+            SCRIPTDBG(7, "Setting power %s property %s to %" PRId64, powername, property, number_value);
         }
     }
     #endif
